Move data sync planning code into its own module
