+++ tests/test_logical_data_sync_planning.py
@@ -1,8 +1,8 @@
 from brad.blueprint.user import UserProvidedBlueprint
 from brad.config.engine import Engine
 from brad.data_sync.logical_plan import ExtractDeltas, TransformDeltas, ApplyDeltas
+from brad.data_sync.planner import make_logical_data_sync_plan
 from brad.planner.data import bootstrap_blueprint
-from brad.planner.data_sync import make_logical_data_sync_plan
 
 
 def test_make_logical_data_sync_plan_simple():
